move validateServiceDefinition to utils package
