+++ packages/api/src/microservice/Qualifier.ts
@@ -1,0 +1,4 @@
+export interface Qualifier {
+  serviceName: string;
+  methodName: string;
+}
