+++ packages/utils/src/qualifier.ts
@@ -1,0 +1,4 @@
+import { MicroserviceApi } from '@scalecube/api';
+
+export const getQualifier = ({ serviceName, methodName }: MicroserviceApi.Qualifier): string =>
+  `${serviceName}/${methodName}`;
